+++ Changelog.md
@@ -1,6 +1,33 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.0.2
+Added a new button in Eltruism > Media, that will automatically apply texture mode to all aura indicators, such as Party/Raid Unitframes  
+Added /eltruism performance command, which will disable several Eltruism and ElvUI settings in order to maximize performance  
+Added another style to BigWigs, Eltruism Top Text, which will show the texts above the progress bar  
+Added an option to change the Time to Arrive on Waypoints to be class colored  
+Added options to disable the Pause/Grey on portrait while the unit is dead  
+Added an option to hide Auras on Nameplates when they are not in combat  
+Added Questie Tracker to the compatibility check  
+Added Lockpicking to the Profession Bag Buttons  
+Added a toggle for Datatext shadows  
+Updated Objective Frame Anchor in Retail to better handle removing it from Edit Mode, this also prevents some taints  
+Updated [eltruism:combatindicator] to use any texture set in the player combat indicator  
+Updated Waypoint Coordinate Inputs to also support TomTom inputs (#Zone X Y)  
+Updated Gradient Backdrop to also work with Offline and Tap Denied states  
+Updated Resting Icon animation to better hide during cinematics  
+Updated target combat indicator to be the same as player  
+Updated power models when the power bar is transparent  
+Updated profile to use Tile3 on the ElvUI Skin  
+Updated Portrait Fix with a few models  
+Updated German locale by Dlarge  
+Updated Auctionator Skin  
+Fixed an issue where entries could be duplicated in the LFG Spec Icons option due to other addons also adding text  
+Fixed an issue where raid control could error when Blizzard Raid frames were enabled  
+Fixed a possible issue with the Ace3 skin where the database was not loaded for tabs  
+Fixed a possible issue with the Teleports Datatext where the item could be invalid  
+Fixed World Quest Icon in the Quest Skin  
+Fixed Shadows when power is detached
 ### 4.0.1
 Added /eltruism autoupdate command, which will run the Eltruism Update function when a new version is installed  
 Added alpha settings to custom color shadows